Replace updated changes

+++ shared/web/public/scripts/storagenodestart.sh
@@ -25,16 +25,23 @@ CONTAINER_NAME=storjlabsSnContainer
 
 echo "$(date)" " Starting Storagenode ${CONTAINER_NAME} ---> " >> "$LOG"
 docker ps -a  >> "$LOG"
-cmd_run="docker run --rm -e SETUP="true"  --mount type=bind,source=${4},destination=/app/identity --mount type=bind,source=${5},destination=/app/config --name ${CONTAINER_NAME} storjlabs/storagenode:latest"
+
 if [[ $# -ge 6 ]]
 then
     cmd="docker run -d --restart unless-stopped --stop-timeout 300 -p ${PORTADDR}:28967 -p ${IPADDR}:14002:14002 -p ${IPADDR}:9000:9000 -e WALLET=${2} -e EMAIL=${6} -e ADDRESS=${1} -e STORAGE=${3}GB --mount type=bind,source=${4},destination=/app/identity --mount type=bind,source=${5},destination=/app/config --name ${CONTAINER_NAME} storjlabs/storagenode:latest "
 else
     cmd="docker run -d --restart unless-stopped --stop-timeout 300 -p ${PORTADDR}:28967 -p ${IPADDR}:14002:14002 -p ${IPADDR}:9000:9000 -e WALLET=${2} -e ADDRESS=${1} -e STORAGE=${3}GB --mount type=bind,source=${4},destination=/app/identity --mount type=bind,source=${5},destination=/app/config --name ${CONTAINER_NAME} storjlabs/storagenode:latest "
 fi
- 
-echo "$cmd_run" >> "$LOG"
-$cmd_run >> "$LOG" 2>&1 
+
+if [ "$(ls -A ${5})" ]; 
+then
+     echo "${5} have data" >> "$LOG"
+else
+    cmd_run="docker run --rm -e SETUP="true"  --mount type=bind,source=${4},destination=/app/identity --mount type=bind,source=${5},destination=/app/config --name ${CONTAINER_NAME} storjlabs/storagenode:latest"
+    echo "$cmd_run" >> "$LOG"
+    $cmd_run >> "$LOG" 2>&1 
+fi
+
 echo "$cmd" >> "$LOG"
 $cmd >> "$LOG" 2>&1 
 echo "$output" >> "$LOG" 
@@ -42,4 +49,3 @@ echo "$output"
 output=$(docker ps -a | grep "${CONTAINER_NAME}" )
 echo "$output" >> "$LOG" 
 cat <<< "$output"
-
